filtering rules by entity classes
